feat: backport training features from Vietnamese-RVC, Applio, and PolTrain
Backports 14 training-related features from 3 community RVC forks.
All changes are additive — no behavior change unless opted in via config.

Vietnamese-RVC (8 features):
- DiscriminatorR AMP safety: cast to float32 before STFT, prevents NaN in fp16
- feature_loss .detach() on real features (saves wasted graph build on G step)
- 'pesto' added to f0 fallback list for DirectML/OpenCL backends
- Configurable f0_min/f0_max from config dict (non-standard vocal ranges)
- Pure-torch mel filter bank (eliminates librosa CPU round-trip, bit-for-bit
  equivalent to librosa.filters.mel(norm='slaney'))
- TF32/compile_all/int8/compile_mode config flags + hardware detection
  (cuda_tf32, cuda_bf16, tf32_support, bf16_support)
- setup_compile() with Triton detection, inductor cache, graph capture opts
- torch.compile embedder when compile_all=True (1.3-2x speedup)

Applio (2 features):
- Vocoder-conditional multiscale_mel_loss auto-enable for RefineGAN/BigVGAN
  (prevents subtle quality degradation from single-scale L1 loss)
- Relative filelist path resolution in TextAudioLoader (makes filelists
  portable across machines)

PolTrain (4 features):
- PolOpt optimizer (Yogi + AdaBelief hybrid with trust-region clamping,
  correct eps placement outside sqrt). Registered as 6th optimizer choice.
  Pairs with CosineAnnealingLR scheduler.
- mel_spectrogram_similarity TensorBoard metric (0-100% interpretable score)
- dB-normalized plot_spectrogram_to_numpy (fixed vmin=-10/vmax=0 range
  makes mel spectrograms comparable across epochs)
- _last.pth final model marker for downstream tooling

New files:
- arvc/engine/models/optimizers/polopt.py (PolOpt optimizer)
- TRAINING_FEATURES_BACKPORTS.md (detailed documentation)

Modified files (11):
- arvc/engine/models/algorithms/discriminators.py
- arvc/engine/models/optimizers/__init__.py
- arvc/engine/training/extract/embedding.py
- arvc/engine/training/extract/feature.py
- arvc/engine/training/runner/data_utils.py
- arvc/engine/training/runner/losses.py
- arvc/engine/training/runner/mel_processing.py
- arvc/engine/training/runner/train.py
- arvc/engine/training/runner/utils.py
- arvc/utils/variables.py

Verification:
- All Python files pass ast.parse syntax check
- PolOpt optimizer: import + step verified
- Pure-torch mel_filter_bank: max diff vs librosa = 7.26e-08 (bit-for-bit)
- mel_spectrogram_similarity: returns valid 0-100% range
- DiscriminatorR: forward pass works with new AMP cast
- Config: all new flags default to safe values (False/None)

Optional config flags (in arvc/configs/config.json):
{
  "tf32": true,              // ~3x matmul speedup on Ampere+ GPUs
  "compile_all": true,       // torch.compile embedder (needs Triton)
  "compile_mode": "default",  // or 'reduce-overhead' / 'max-autotune'
  "int8": false,             // 8-bit optimizer (needs bitsandbytes)
  "f0_min": 50,              // customize for non-standard voices
  "f0_max": 1100
}

Sources:
- https://github.com/PhamHuynhAnh16/Vietnamese-RVC
- https://github.com/IAHispano/Applio
- https://github.com/Politrees/PolTrain

+++ arvc/engine/models/algorithms/discriminators.py
@@ -129,7 +129,11 @@ class DiscriminatorR(torch.nn.Module):
 
     def forward(self, x):
         fmap = []
-        x = self.spectrogram(x).unsqueeze(1)
+        # AMP safety (from Vietnamese-RVC): cast to float32 before STFT to
+        # avoid NaN/Inf in fp16 autocast. torch.stft is not numerically
+        # stable in fp16 on all backends — this matches the universal pattern
+        # used by DiscriminatorP and the main mel loss.
+        x = self.spectrogram(x.float()).unsqueeze(1).to(x.dtype)
         
         for layer in self.convs:
             x = F.leaky_relu(layer(x), self.lrelu_slope)
